Merge master

+++ src/Resources/contao/classes/RealEstateImporter.php
@@ -458,7 +458,7 @@ class RealEstateImporter extends \BackendModule
                         // Skip if condition dont match
                         if ($interfaceMapping->oiConditionField && $interfaceMapping->oiConditionValue)
                         {
-                            if ($interfaceMapping->oiConditionValue !== $this->getFieldData($interfaceMapping->oiConditionField, $group))
+                            if (!$this->isValidCondition($interfaceMapping->oiConditionValue, $this->getFieldData($interfaceMapping->oiConditionField, $group)))
                             {
                                 if ($interfaceMapping->forceActive)
                                 {
@@ -551,6 +551,36 @@ class RealEstateImporter extends \BackendModule
         }
 
         return $this->updateCatalog($contactPersonRecords, $realEstateRecords);
+    }
+
+    /**
+     * Checks if the condition is valid
+     *
+     * @param $strCondition
+     * @param $fieldValue
+     *
+     * @return bool
+     */
+    protected function isValidCondition($strCondition, $fieldValue)
+    {
+        $arrConditionValues = explode('|', $strCondition);
+
+        if(count($arrConditionValues) > 1)
+        {
+            foreach ($arrConditionValues as $conditionValue)
+            {
+                if($conditionValue === $fieldValue)
+                {
+                    return true;
+                }
+            }
+
+            return false;
+        }
+        else
+        {
+            return $strCondition === $fieldValue;
+        }
     }
 
     /**
@@ -821,7 +851,7 @@ class RealEstateImporter extends \BackendModule
         */
 
         try {
-            $this->data = simplexml_load_string($data);
+            $this->data = simplexml_load_string($data, 'SimpleXMLElement', LIBXML_NOCDATA);
         } catch (\Exception $e) {
             return false;
         }
@@ -1230,6 +1260,12 @@ class RealEstateImporter extends \BackendModule
         $objFilesFolder = $interfaceMapping->type === 'tl_contact_person' ? $this->objFilesFolderContactPerson : $this->objFilesFolder;
 
         $check = next($tmpGroup->check);
+        $isMd5 = false;
+
+        if (isset($check) && FilesHelper::isValidMd5($check))
+        {
+            $isMd5 = true;
+        }
 
         // HOOK: add custom logic
         if (isset($GLOBALS['TL_HOOKS']['realEstateImportSaveImage']) && \is_array($GLOBALS['TL_HOOKS']['realEstateImportSaveImage']))
@@ -1255,7 +1291,7 @@ class RealEstateImporter extends \BackendModule
         $filePath = $objFilesFolder->path . '/' . $this->uniqueProviderValue . '/' . ($interfaceMapping->type === 'tl_real_estate' ? $this->uniqueValue . '/' : '') . $value;
         $existingFile = \FilesModel::findByPath($filePath);
 
-        if ($existingFile !== null && $existingFile->hash === $check)
+        if ($existingFile !== null && $isMd5 && $existingFile->hash === $check)
         {
             $values[] = $existingFile->uuid;
             //$this->addLog('Skip image: ' . ($existingFile->hash === $check ? 'Image already exists and has not changed' : 'Image does not exist'), 3, 'info', array(
@@ -1274,7 +1310,7 @@ class RealEstateImporter extends \BackendModule
         }
 
         // Delete file, if hash dont match
-        if ($check !== false && $objFile->hash !== $check)
+        if ($check !== false && $isMd5 && $objFile->hash !== $check)
         {
             $objFile->delete();
             return false;
